Do not loose uploaded photos in case of imgur failure
If the imgur upload fails, the exception is handled so that the dive
is anyways updated with the previous uploaded photos within the same
task.

Also, it will keep trying to add the subsequent photos.

A task should run periodically to see if there are leftovers in the
blobstore.

It should also schedule another task to retry the failed uploads,
and notify an administrator that something might be going on.