Added generic http support module
It provides a Connection class that can be used with "with".

It is refactored from the imgur module.

It also provides a function to determine wether an imgur link
is "good" or not.

Good means that the link exists and that it doesn't point to that
stupid 41bytes gif that someone had fun uploading.